Tease out github_services.py
